feat: Separate time control component.
