+++ components/ADempiere/Form/index.vue
@@ -9,11 +9,11 @@
 
  This program is distributed in the hope that it will be useful,
  but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  GNU General Public License for more details.
 
  You should have received a copy of the GNU General Public License
- along with this program.  If not, see <https:www.gnu.org/licenses/>.
+ along with this program. If not, see <https:www.gnu.org/licenses/>.
 -->
 <template>
   <component
@@ -83,7 +83,6 @@ export default {
           form = import('@theme/components/ADempiere/Form/TimeControl')
           break
         case 'TimeRecord':
-        // case 'org.spin.tcs.form.TimeControl':
           form = import('@theme/components/ADempiere/Form/TimeRecord')
           break
         case 'VPayPrint':
